[DONE] Feature import video v1 (#888)
* Add import_video module

* Remove external recordings from Meeting module and this functionnality in Import_video module

* Add USE_IMPORT_VIDEO and RESTRICT_EDIT_IMPORT_VIDEO_ACCESS_TO_STAFF_ONLY parameters

* Manage translations for the Import_video module

* Display the new Import_video module (as a 'My external videos' link), if necessary

* Manage the USE_IMPORT_VIDEO and RESTRICT_EDIT_IMPORT_VIDEO_ACCESS_TO_STAFF_ONLY parameters to display, or not, the new Import_video module

* Add USE_IMPORT_VIDEO parameter to test this module

* Add the secure_post_request function

* Modify to remove external recordings management

* Refactor BBB internal and external recordings source code

* Modify to remove external recordings management and refactor internal recordings source code

* Add new Import_video module

* Add URLs from the new Import_video module

* get shared secret for bbb test setting

* It's a Python module

* Rename Recording into ExternalRecording for Import_video module

* Add an Import column to the external recordings table

* New translations for the import_video module

* Default value (True) for USE_IMPORT_VIDEO and RESTRICT_EDIT_IMPORT_VIDEO_ACCESS_TO_STAFF_ONLY parameters

* Add OwnerWidget and AddOwnerWidget classes (refactoring source code)

* Rename Recording into InternalRecording for Meeting module

* use beautifulsoup to get shared secret value for bbb

* Refactor source code from meeting and import_video modules

* Change comments

* Change Meta : Recording to External Recording

* Change translations for import_video module

* Remove USE_IMPORT_VIDEO and RESTRICT_EDIT_IMPORT_VIDEO_ACCESS_TO_STAFF_ONLY parameters

* Rename import_video menu in the navigation bar

* Add reference to import_video/context_processors.py

* Create a context_processors.py to the import_video module, with USE_IMPORT_VIDEO and RESTRICT_EDIT_IMPORT_VIDEO_ACCESS_TO_STAFF_ONLY parameters

* Remove useless delete.html page

* Correction of a mistake

* Correction of a mistake

* Translations due to a correction of a mistake

---------
