+++ pod/import_video/static/css/import_video.css
@@ -1,0 +1,109 @@
+#recordings.table {
+  --bs-table-color: var(--pod-font-color);
+  --bs-table-striped-color: var(--pod-font-color);
+}
+
+
+/* Loading while uploading */
+.recording-loader {
+  position: absolute;
+  left: 50%;
+  top: 50%;
+  z-index: 1;
+  width: 350px;
+  height: 160px;
+  background-color: var(--pod-background);
+  border: 1px solid var(--pod-primary);
+  text-align: center;
+  color: var(--pod-font-color);
+  padding: 10px;
+  display: none;
+}
+.recording-loading {
+  margin: auto;
+  border: 6px solid #f3f3f3;
+  border-top: 6px solid var(--pod-primary);
+  border-radius: 50%;
+  width: 50px;
+  height: 50px;
+  animation: recording-spin 3s linear infinite;      
+}
+
+@-webkit-keyframes recording-spin {
+  0% { -webkit-transform: rotate(0deg); }
+  100% { -webkit-transform: rotate(360deg); }
+}
+
+@keyframes recording-spin {
+  0% { transform: rotate(0deg); }
+  100% { transform: rotate(360deg); }
+}
+/* */
+/* Message error */
+div.alert {
+  border-radius: 6px;
+  display: table;
+  width: 100%;
+  padding-left: 78px;
+  position: relative;
+  padding-right: 60px;
+  border: 1px solid #e6e6e6;
+  background: #FFF;
+}
+
+div.alert .icon {
+  text-align: center;
+  width: 58px;
+  height: 100%;
+  position: absolute;
+  top: 0;
+  left: 0;
+  border: 1px solid #bdbdbd;
+  padding-top: 15px;
+  border-radius: 6px 0 0 6px;
+}
+
+div.alert .icon i {
+  font-size: 20px;
+  color: #fff;
+  left: 21px;
+  margin-top: -15px;
+  position: absolute;
+  top: 50%;
+}
+
+div.alert.alert-success .icon,
+div.alert.alert-success .icon:after {
+  border-color: none;
+  background: #00986a;
+}
+
+div.alert.alert-info .icon,
+div.alert.alert-info .icon:after {
+  border-color: none;
+  background: #00b3c8;
+}
+
+div.alert.alert-warning .icon,
+div.alert.alert-warning .icon:after {
+  border: none;
+  background: #f9af2c;
+}
+
+div.alert.alert-error .icon,
+div.alert.alert-error .icon:after {
+  border-color: none;
+  background: #c82630;
+}
+
+div.alert .proposition {
+  color: #000;
+  font-weight: Bold;
+}
+div.alert .proposition::before {
+  content: "\F137";
+  margin-right: 15px;
+  font-family: bootstrap-icons;
+  vertical-align: -.125em;
+}
+/* */
