+++ pod/main/templates/navbar.html
@@ -204,6 +204,10 @@
               {% if USE_OPENCAST_STUDIO %}
                 <a class="nav-item nav-link" href="{% url 'recorder:studio_pod'%}"><i class="bi bi-camera-video pod-nav-link-icon mx-1" aria-hidden="true"></i>{% trans "Video Record" %}</a>
               {% endif %}
+              {% show_import_video_button as import_video_button %}
+              {% if import_video_button %}
+                <a class="nav-item nav-link" href="{% url 'import_video:external_recordings'%}"><i class="bi bi-collection-play-fill pod-nav-link-icon mx-1" aria-hidden="true"></i>{% trans "Import an external video" %}</a>
+              {% endif %}
               {% if "live" in THIRD_PARTY_APPS and request.user|can_manage_event_filter %}
                 <a class="nav-item nav-link" href="{% url 'live:my_events' %}" title="{% trans 'My events' %}">
                   <i class="bi bi-broadcast pod-nav-link-icon mx-1" aria-hidden="true"></i>{% trans 'My events' %}
