+++ pod/main/utils.py
@@ -1,6 +1,8 @@
 """Useful functions for the entire Pod application."""
 from django.contrib import messages
+from django.core.exceptions import PermissionDenied
 from django.utils.html import mark_safe
+from django.utils.translation import gettext_lazy as _
 
 
 def is_ajax(request) -> bool:
@@ -28,3 +30,19 @@ def display_message_with_icon(request, type, message):
     msg = "<div class='icon'><i class='bi bi-" + icon + "'></i></div>"
     msg += message
     messages.add_message(request, type, mark_safe(msg))
+
+
+def secure_post_request(request):
+    """Secure that this request is POST type.
+
+    Args:
+        request (Request): HTTP request
+
+    Raises:
+        PermissionDenied: if method not POST
+    """
+    if request.method != "POST":
+        display_message_with_icon(
+            request, messages.WARNING, _("This view cannot be accessed directly.")
+        )
+        raise PermissionDenied
