+++ pod/meeting/admin.py
@@ -1,3 +1,4 @@
+"""Admin for Meeting module."""
 from django.contrib import admin
 from django.contrib.sites.shortcuts import get_current_site
 from django.urls import reverse
@@ -5,7 +6,7 @@ from django.utils.translation import ugettext_lazy as _
 from django.utils.html import mark_safe
 from django.contrib.admin import widgets
 
-from .models import Meeting, Recording
+from .models import Meeting, InternalRecording
 from .forms import (
     MeetingForm,
     MEETING_MAIN_FIELDS,
@@ -166,9 +167,9 @@ class MeetingAdmin(admin.ModelAdmin):
         return qs
 
 
-@admin.register(Recording)
-class RecordingAdmin(admin.ModelAdmin):
-    """Administration for internal and external recordings.
+@admin.register(InternalRecording)
+class InternalRecordingAdmin(admin.ModelAdmin):
+    """Administration for BBB internal recordings.
 
     Args:
         admin (ModelAdmin): admin model
@@ -177,10 +178,8 @@ class RecordingAdmin(admin.ModelAdmin):
     list_display = (
         "name",
         "start_at",
-        "is_internal",
         "recording_id",
         "meeting",
-        "type",
         "owner",
     )
     search_fields = [
