+++ pod/meeting/tests/test_models.py
@@ -1,13 +1,13 @@
-from django.test import TestCase
-from django.contrib.auth.models import User
-from django.template.defaultfilters import slugify
-from datetime import datetime, date
-from django.core.exceptions import ValidationError
-from django.contrib.sites.models import Site
-
+"""Tests the models for meeting module."""
 import random
 
-from ..models import Meeting, Recording
+from ..models import Meeting, InternalRecording
+from datetime import datetime, date
+from django.contrib.auth.models import User
+from django.contrib.sites.models import Site
+from django.core.exceptions import ValidationError
+from django.test import TestCase
+from django.template.defaultfilters import slugify
 from pod.authentication.models import AccessGroup
 
 
@@ -725,8 +725,8 @@ class OccurencesMeetingTestCase(TestCase):
         )
 
 
-class RecordingTestCase(TestCase):
-    """List of recordings model tests, internal or external.
+class InternalRecordingTestCase(TestCase):
+    """List of recordings model tests, only internal.
 
     Args:
         TestCase (class): test case
@@ -748,19 +748,17 @@ class RecordingTestCase(TestCase):
         )
         meeting2.additional_owners.add(user1)
         meeting2.additional_owners.add(user2)
-        Recording.objects.create(
+        InternalRecording.objects.create(
             id=1,
             name="test recording1",
-            is_internal=True,
             owner=user,
             recording_id="c057c39d3dc59d9e9516d95f76eb",
             meeting=meeting1,
             site=Site.objects.get(id=1),
         )
-        Recording.objects.create(
+        InternalRecording.objects.create(
             id=2,
             name="test recording2",
-            is_internal=True,
             owner=user,
             recording_id="d058c39d3dc59d9e9516d95f76eb",
             meeting=meeting2,
@@ -768,19 +766,10 @@ class RecordingTestCase(TestCase):
             uploaded_to_pod_by=user2,
             site=Site.objects.get(id=1),
         )
-        Recording.objects.create(
-            id=3,
-            name="test recording3",
-            is_internal=False,
-            owner=user,
-            start_at=datetime(2022, 4, 24, 14, 0, 0),
-            site=Site.objects.get(id=1),
-            type="bigbluebutton",
-        )
 
     def test_default_attributs(self):
         """Check all default attributs values when creating a recording."""
-        recordings = Recording.objects.all()
+        recordings = InternalRecording.objects.all()
         self.assertGreaterEqual(
             recordings[0].start_at.date(), recordings[1].start_at.date()
         )
@@ -788,26 +777,25 @@ class RecordingTestCase(TestCase):
     def test_with_attributs(self):
         """Check all attributs values passed when creating a recording."""
         meeting2 = Meeting.objects.get(id=2)
-        recording2 = Recording.objects.get(id=2)
+        recording2 = InternalRecording.objects.get(id=2)
         self.assertEqual(recording2.name, "test recording2")
         user2 = User.objects.get(username="pod2")
         self.assertEqual(recording2.uploaded_to_pod_by, user2)
-        self.assertEqual(recording2.is_internal, True)
         self.assertEqual(recording2.meeting, meeting2)
 
     def test_change_attributs(self):
         """Change attributs values in a recording and save it."""
-        recording1 = Recording.objects.get(id=1)
+        recording1 = InternalRecording.objects.get(id=1)
         self.assertEqual(recording1.name, "test recording1")
         self.assertEqual(recording1.recording_id, "c057c39d3dc59d9e9516d95f76eb")
         recording1.name = "New test recording1 !"
         recording1.recording_id = "d058c39d3dc59d9e9516d95f76eb"
         recording1.save()
-        newrecording1 = Recording.objects.get(id=1)
+        newrecording1 = InternalRecording.objects.get(id=1)
         self.assertEqual(newrecording1.name, "New test recording1 !")
         self.assertEqual(newrecording1.recording_id, "d058c39d3dc59d9e9516d95f76eb")
 
     def test_delete_object(self):
         """Delete a recording."""
-        Recording.objects.filter(name="test recording2").delete()
-        self.assertEqual(Recording.objects.all().count(), 2)
+        InternalRecording.objects.filter(name="test recording2").delete()
+        self.assertEqual(InternalRecording.objects.all().count(), 1)
