+++ pod/meeting/urls.py
@@ -1,3 +1,4 @@
+"""URLs for Meeting module."""
 from django.urls import path
 
 from . import views
@@ -33,32 +34,9 @@ if not views.MEETING_DISABLE_RECORD:
             name="upload_internal_recording_to_pod",
         ),
         path(
-            "upload_recording_to_pod/<slug:record_id>/",
-            views.upload_external_recording_to_pod,
-            name="upload_external_recording_to_pod",
-        ),
-        path(
             "delete_recording/<slug:meeting_id>/<slug:recording_id>/",
             views.delete_internal_recording,
             name="delete_internal_recording",
-        ),
-        path(
-            "external_recordings/", views.external_recordings, name="external_recordings"
-        ),
-        path(
-            "external_recording/add",
-            views.add_or_edit_external_recording,
-            name="add_external_recording",
-        ),
-        path(
-            "external_recording/edit/<slug:id>/",
-            views.add_or_edit_external_recording,
-            name="edit_external_recording",
-        ),
-        path(
-            "external_recording/delete/<slug:id>/",
-            views.delete_external_recording,
-            name="delete_external_recording",
         ),
     ]
