+++ pod/meeting/views.py
@@ -1,48 +1,34 @@
 """Views of the Meeting module."""
 import os
 import bleach
-import requests
 
-# from django.utils.dateparse import parse_duration
-# from datetime import datetime as dt
-
-from django.shortcuts import render, redirect
-from django.http import JsonResponse, HttpResponse
-from django.contrib.sites.shortcuts import get_current_site
-from django.templatetags.static import static
-from django.contrib.auth.decorators import login_required
-from django.utils.translation import gettext_lazy as _
-from django.utils.html import mark_safe
-from django.shortcuts import get_object_or_404
-from django.core.exceptions import SuspiciousOperation
-from django.views.decorators.csrf import csrf_protect
-from django.core.exceptions import PermissionDenied
-from django.contrib import messages
-from django.urls import reverse
-from django.views.decorators.csrf import ensure_csrf_cookie
-from django.conf import settings
-from django.core.mail import EmailMultiAlternatives
-from django.utils import timezone
-from datetime import datetime
-
-from html.parser import HTMLParser
-import shutil
-from pod.video.models import Video
-from pod.video.models import Type
-
-from .models import Meeting, StatelessRecording, Recording
-from .utils import get_nth_week_number
 from .forms import MeetingForm, MeetingDeleteForm, MeetingPasswordForm
-from .forms import MeetingInviteForm, RecordingForm
+from .forms import MeetingInviteForm
+from .models import Meeting, InternalRecording
+from .utils import get_nth_week_number
+from datetime import datetime
+from django.conf import settings
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.contrib.sites.shortcuts import get_current_site
+from django.core.exceptions import SuspiciousOperation
+from django.core.exceptions import PermissionDenied
+from django.core.mail import EmailMultiAlternatives
+from django.http import JsonResponse, HttpResponse
+from django.shortcuts import get_object_or_404
+from django.shortcuts import render, redirect
+from django.templatetags.static import static
+from django.urls import reverse
+from django.utils import timezone
+from django.utils.html import mark_safe
+from django.utils.translation import gettext_lazy as _
+from django.views.decorators.csrf import csrf_protect
+from django.views.decorators.csrf import ensure_csrf_cookie
+from pod.import_video.utils import StatelessRecording
+from pod.import_video.utils import secure_request_for_upload, parse_remote_file
+from pod.import_video.utils import download_video_file, save_video
 from pod.main.views import in_maintenance, TEMPLATE_VISIBLE_SETTINGS
-from pod.main.utils import display_message_with_icon
-
-# For Youtube download
-from pytube import YouTube
-from pytube.exceptions import PytubeError, VideoUnavailable
-
-# For PeerTube download
-import json
+from pod.main.utils import secure_post_request, display_message_with_icon
 
 RESTRICT_EDIT_MEETING_ACCESS_TO_STAFF_ONLY = getattr(
     settings, "RESTRICT_EDIT_MEETING_ACCESS_TO_STAFF_ONLY", False
@@ -68,8 +54,6 @@ BBB_MEETING_INFO = getattr(
     },
 )
 MEETING_DISABLE_RECORD = getattr(settings, "MEETING_DISABLE_RECORD", True)
-
-DEFAULT_TYPE_ID = getattr(settings, "DEFAULT_TYPE_ID", 1)
 
 VIDEO_ALLOWED_EXTENSIONS = getattr(
     settings,
@@ -106,6 +90,7 @@ __TITLE_SITE__ = (
 
 @login_required(redirect_field_name="referrer")
 def my_meetings(request):
+    """List the meetings."""
     site = get_current_site(request)
     if RESTRICT_EDIT_MEETING_ACCESS_TO_STAFF_ONLY and request.user.is_staff is False:
         return render(request, "meeting/my_meetings.html", {"access_not_allowed": True})
@@ -137,6 +122,7 @@ def my_meetings(request):
 @ensure_csrf_cookie
 @login_required(redirect_field_name="referrer")
 def add_or_edit(request, meeting_id=None):
+    """Add or edit a meeting."""
     if in_maintenance():
         return redirect(reverse("maintenance"))
     meeting = (
@@ -206,6 +192,7 @@ def add_or_edit(request, meeting_id=None):
 
 
 def save_meeting_form(request, form):
+    """Save a meeting form."""
     meeting = form.save(commit=False)
     meeting.site = get_current_site(request)
     if (
@@ -226,6 +213,7 @@ def save_meeting_form(request, form):
 @ensure_csrf_cookie
 @login_required(redirect_field_name="referrer")
 def delete(request, meeting_id):
+    """Delete a meeting."""
     meeting = get_object_or_404(
         Meeting, meeting_id=meeting_id, site=get_current_site(request)
     )
@@ -261,6 +249,7 @@ def delete(request, meeting_id):
 @csrf_protect
 @ensure_csrf_cookie
 def join(request, meeting_id, direct_access=None):
+    """Join a meeting."""
     try:
         id = int(meeting_id[: meeting_id.find("-")])
     except ValueError:
@@ -288,6 +277,7 @@ def join(request, meeting_id, direct_access=None):
 
 
 def render_show_page(request, meeting, show_page, direct_access):
+    """Render show page."""
     if show_page and direct_access and request.user.is_authenticated:
         # join as attendee
         # get user name and redirect to BBB
@@ -313,6 +303,7 @@ def render_show_page(request, meeting, show_page, direct_access):
 
 
 def join_as_moderator(request, meeting):
+    """Join as a moderator."""
     try:
         created = meeting.create(request)
         if created:
@@ -348,6 +339,7 @@ def join_as_moderator(request, meeting):
 
 
 def check_user(request):
+    """Check user."""
     if request.user.is_authenticated:
         display_message_with_icon(
             request, messages.ERROR, _("You cannot access to this meeting.")
@@ -358,6 +350,7 @@ def check_user(request):
 
 
 def check_form(request, meeting, remove_password_in_form):
+    """Check form."""
     current_user = request.user if request.user.is_authenticated else None
     form = MeetingPasswordForm(
         current_user=current_user,
@@ -410,6 +403,7 @@ def check_form(request, meeting, remove_password_in_form):
 
 
 def is_in_meeting_groups(user, meeting):
+    """Return if user in the meeting."""
     return user.owner.accessgroup_set.filter(
         code_name__in=[
             name[0] for name in meeting.restrict_access_to_groups.values_list("code_name")
@@ -440,6 +434,7 @@ def get_meeting_access(request, meeting):
 @ensure_csrf_cookie
 # @login_required(redirect_field_name="referrer")
 def status(request, meeting_id):
+    """Status of a meeting, in JSON format."""
     meeting = get_object_or_404(
         Meeting, meeting_id=meeting_id, site=get_current_site(request)
     )
@@ -460,6 +455,7 @@ def status(request, meeting_id):
 @ensure_csrf_cookie
 @login_required(redirect_field_name="referrer")
 def end(request, meeting_id):
+    """End meeting, in JSON format.."""
     meeting = get_object_or_404(
         Meeting, meeting_id=meeting_id, site=get_current_site(request)
     )
@@ -492,6 +488,7 @@ def end(request, meeting_id):
 @ensure_csrf_cookie
 @login_required(redirect_field_name="referrer")
 def get_meeting_info(request, meeting_id):
+    """Get meeting info, in JSON format."""
     meeting = get_object_or_404(
         Meeting, meeting_id=meeting_id, site=get_current_site(request)
     )
@@ -543,7 +540,7 @@ def internal_recordings(request, meeting_id):
     # Secure the list of internal recordings
     secure_internal_recordings(request, meeting)
 
-    # Additional owners can't delete these recordings
+    # The user can delete this recording ?
     can_delete = get_can_delete_recordings(request, meeting)
 
     meeting_recordings = meeting.get_recordings()
@@ -571,8 +568,8 @@ def internal_recordings(request, meeting_id):
             # Only the owner can delete their recordings
             bbb_recording.canDelete = can_delete
             # Search for more informations from database (if already uploaded to Pod)
-            recording = Recording.objects.filter(
-                recording_id=data["recordID"], is_internal=True
+            recording = InternalRecording.objects.filter(
+                recording_id=data["recordID"]
             ).first()
             if recording:
                 bbb_recording.uploadedToPodBy = recording.uploaded_to_pod_by
@@ -587,22 +584,6 @@ def internal_recordings(request, meeting_id):
             "page_title": _("Meeting recordings"),
         },
     )
-
-
-def secure_post_request(request):
-    """Secure that this request is POST type.
-
-    Args:
-        request (Request): HTTP request
-
-    Raises:
-        PermissionDenied: if method not POST
-    """
-    if request.method != "POST":
-        display_message_with_icon(
-            request, messages.WARNING, _("This view cannot be accessed directly.")
-        )
-        raise PermissionDenied
 
 
 def secure_internal_recordings(request, meeting):
@@ -621,56 +602,11 @@ def secure_internal_recordings(request, meeting):
         and not (
             request.user.is_superuser or request.user.has_perm("meeting.view_meeting")
         )
-        and (request.user not in meeting.additional_owners.all())
     ):
         display_message_with_icon(
             request, messages.ERROR, _("You cannot view the recordings of this meeting.")
         )
         raise PermissionDenied
-
-
-def secure_external_recording(request, recording):
-    """Secure an external recording.
-
-    Args:
-        request (Request): HTTP request
-        recording (Recording): Recording instance
-
-    Raises:
-        PermissionDenied: if user not allowed
-    """
-    if (
-        recording
-        and request.user != recording.owner
-        and not (
-            request.user.is_superuser or request.user.has_perm("meeting.view_recording")
-        )
-        and (request.user not in recording.additional_owners.all())
-    ):
-        display_message_with_icon(
-            request, messages.ERROR, _("You cannot view this recording.")
-        )
-        raise PermissionDenied
-
-
-def secure_request_for_upload(request):
-    """Check that the request is correct for uploading a recording.
-
-    Args:
-        request (Request): HTTP request
-
-    Raises:
-        ValueError: if bad data
-    """
-    # Source_url and recording_name are necessary
-    if request.POST.get("source_url") == "" or request.POST.get("recording_name") == "":
-        msg = {}
-        msg["error"] = _("Impossible to upload to Pod the video")
-        msg["message"] = _("No URL found / No recording name")
-        msg["proposition"] = _(
-            "Try changing the record type or address for this recording"
-        )
-        raise ValueError(msg)
 
 
 def get_can_delete_recordings(request, meeting):
@@ -685,25 +621,10 @@ def get_can_delete_recordings(request, meeting):
     """
     can_delete = False
 
-    # Additional owners can't delete these recordings
     if (
         request.user == meeting.owner
         or (request.user.is_superuser)
         or (request.user.has_perm("meeting.delete_meeting"))
-    ):
-        can_delete = True
-    return can_delete
-
-
-def get_can_delete_external_recording(request, owner):
-    """Return True if current user can delete this recording."""
-    can_delete = False
-
-    # Only owner can delete this external recording
-    if (
-        request.user == owner
-        or (request.user.is_superuser)
-        or (request.user.has_perm("meeting.delete_external_recording"))
     ):
         can_delete = True
     return can_delete
@@ -756,6 +677,7 @@ def delete_internal_recording(request, meeting_id, recording_id):
 
 
 def get_meeting_info_json(info):
+    """Get meeting info in JSON format."""
     response = {}
     for key in info:
         temp_key = key.split("__")[0] if "__" in key else key
@@ -771,6 +693,7 @@ def get_meeting_info_json(info):
 
 
 def end_callback(request, meeting_id):
+    """End the BBB callback."""
     meeting = get_object_or_404(
         Meeting, meeting_id=meeting_id, site=get_current_site(request)
     )
@@ -783,6 +706,7 @@ def end_callback(request, meeting_id):
 @ensure_csrf_cookie
 @login_required(redirect_field_name="referrer")
 def invite(request, meeting_id):
+    """Invite users to a BBB meeting."""
     meeting = get_object_or_404(
         Meeting, meeting_id=meeting_id, site=get_current_site(request)
     )
@@ -813,6 +737,7 @@ def invite(request, meeting_id):
 
 
 def get_dest_emails(meeting, form):
+    """Recipient emails."""
     emails = form.cleaned_data["emails"]
     if form.cleaned_data["owner_copy"] is True:
         emails.append(meeting.owner.email)
@@ -822,6 +747,7 @@ def get_dest_emails(meeting, form):
 
 
 def send_invite(request, meeting, emails):
+    """Send invitations to users."""
     subject = _("%(owner)s invites you to the meeting %(meeting_title)s") % {
         "owner": meeting.owner.get_full_name(),
         "meeting_title": meeting.name,
@@ -844,6 +770,7 @@ def send_invite(request, meeting, emails):
 
 
 def get_html_content(request, meeting):
+    """Get HTML format content."""
     join_link = request.build_absolute_uri(
         reverse("meeting:join", args=(meeting.meeting_id,))
     )
@@ -909,6 +836,7 @@ def get_html_content(request, meeting):
 
 
 def create_ics(request, meeting):
+    """Create ICS format."""
     join_link = request.build_absolute_uri(
         reverse("meeting:join", args=(meeting.meeting_id,))
     )
@@ -975,7 +903,8 @@ def create_ics(request, meeting):
 
 
 def get_rrule(meeting):
-    """
+    """Get recurrence rule.
+
     i.e:
     RRULE:FREQ=DAILY;INTERVAL=2;COUNT=28
     RRULE:FREQ=WEEKLY;INTERVAL=2;BYDAY=MO,TU;UNTIL=20221011T100000Z
@@ -1041,11 +970,10 @@ def upload_internal_recording_to_pod(request, recording_id, meeting_id):
         save_internal_recording(request, recording_id, recording_title, meeting_id)
         # Get it
         recording = get_object_or_404(
-            Recording,
+            InternalRecording,
             recording_id=recording_id,
             meeting_id=meeting.id,
             site=get_current_site(request),
-            is_internal=True,
         )
 
         # Upload this recording
@@ -1070,415 +998,7 @@ def upload_internal_recording_to_pod(request, recording_id, meeting_id):
     return redirect(reverse("meeting:internal_recordings", args=(meeting.meeting_id,)))
 
 
-@csrf_protect
-@ensure_csrf_cookie
-@login_required(redirect_field_name="referrer")
-def upload_external_recording_to_pod(request, record_id):
-    """Upload external recording to Pod.
-
-    Args:
-        request (Request): HTTP request
-        recording_id (Integer): record id (in database)
-
-    Raises:
-        PermissionDenied: if user not allowed
-
-    Returns:
-        HTTP Response: Redirect to the external recordings list
-    """
-    recording = get_object_or_404(Recording, id=record_id, site=get_current_site(request))
-
-    # Secure this external recording
-    secure_external_recording(request, recording)
-
-    # Only POST request
-    secure_post_request(request)
-
-    msg = ""
-    upload = False
-    try:
-        upload = upload_recording_to_pod(request, record_id)
-    except ValueError as ve:
-        args = ve.args[0]
-        msg = ""
-        if args.get("error"):
-            msg += "<strong>%s</strong><br>" % (args["error"])
-        if args.get("message"):
-            msg += args["message"]
-        if args.get("proposition"):
-            msg += "<br><span class='proposition'>%s</span>" % (args["proposition"])
-    if upload and msg == "":
-        msg += _(
-            "The recording has been uploaded to Pod. "
-            "You can see the generated video in My videos."
-        )
-        display_message_with_icon(request, messages.INFO, msg)
-    else:
-        display_message_with_icon(request, messages.ERROR, msg)
-    return redirect(reverse("meeting:external_recordings", args=()))
-
-
-@login_required(redirect_field_name="referrer")
-def external_recordings(request):
-    """List external recordings.
-
-    Args:
-        request (Request): HTTP Request
-
-    Returns:
-        HTTPResponse: external recordings list
-    """
-    # print(timezone.localtime().strftime("%y%m%d-%H%M%S"))
-
-    if RESTRICT_EDIT_MEETING_ACCESS_TO_STAFF_ONLY and request.user.is_staff is False:
-        return render(
-            request, "meeting/external_recordings.html", {"access_not_allowed": True}
-        )
-
-    site = get_current_site(request)
-
-    # List of the external recordings from the database
-    external_recordings = Recording.objects.filter(
-        owner__id=request.user.id, is_internal=False, site=site
-    ) | request.user.owners_recordings.all().filter(site=site)
-    external_recordings = external_recordings.order_by("-id").distinct()
-
-    recordings = []
-    for data in external_recordings:
-        # Management of the stateless recording
-        recording = StatelessRecording(data.id, data.name, "published")
-        # Upload to Pod is always possible in such a case
-        recording.canUpload = True
-        # Only owner can delete this external recording
-        recording.canDelete = get_can_delete_external_recording(request, data.owner)
-
-        recording.startTime = data.start_at
-        # recording.endTime = data.endTime
-
-        # Management of the external recording type
-        if data.type == "bigbluebutton":
-            # For BBB, external URL can be the video or presentation playback
-            if data.source_url.find("playback/video") != -1:
-                # Management for standards video URLs with BBB or Scalelite server
-                recording.videoUrl = data.source_url
-            elif data.source_url.find("playback/presentation/2.3") != -1:
-                # Management for standards presentation URLs with BBB or Scalelite server
-                # Add computed video playback
-                recording.videoUrl = data.source_url.replace(
-                    "playback/presentation/2.3", "playback/video"
-                )
-            else:
-                # Management of other situations, non standards URLs
-                recording.videoUrl = data.source_url
-        else:
-            # For PeerTube, Video file, Youtube
-            recording.videoUrl = data.source_url
-
-        # Display type label
-        recording.type = data.get_type_display
-
-        recording.uploadedToPodBy = data.uploaded_to_pod_by
-
-        recordings.append(recording)
-
-    return render(
-        request,
-        "meeting/external_recordings.html",
-        {
-            "recordings": recordings,
-            "page_title": _("External recordings"),
-        },
-    )
-
-
-@csrf_protect
-@ensure_csrf_cookie
-@login_required(redirect_field_name="referrer")
-def add_or_edit_external_recording(request, id=None):
-    """Add or edit an external recording.
-
-    Args:
-        request (Request): HTTP request
-        id (Integer, optional): external recording id. Defaults to None (for creation).
-
-    Raises:
-        PermissionDenied: if user not allowed
-
-    Returns:
-        HTTPResponse: edition page
-    """
-    if in_maintenance():
-        return redirect(reverse("maintenance"))
-
-    recording = (
-        get_object_or_404(Recording, id=id, site=get_current_site(request))
-        if id
-        else None
-    )
-
-    # Secure external recording
-    secure_external_recording(request, recording)
-
-    if RESTRICT_EDIT_MEETING_ACCESS_TO_STAFF_ONLY and request.user.is_staff is False:
-        return render(
-            request,
-            "meeting/add_or_edit_external_recording.html",
-            {"access_not_allowed": True},
-        )
-
-    default_owner = recording.owner.pk if recording else request.user.pk
-    form = RecordingForm(
-        instance=recording,
-        is_staff=request.user.is_staff,
-        is_superuser=request.user.is_superuser,
-        current_user=request.user,
-        initial={"owner": default_owner, "id": id},
-    )
-
-    if request.method == "POST":
-        form = RecordingForm(
-            request.POST,
-            instance=recording,
-            is_staff=request.user.is_staff,
-            is_superuser=request.user.is_superuser,
-            current_user=request.user,
-            current_lang=request.LANGUAGE_CODE,
-        )
-        if form.is_valid():
-            recording = save_recording_form(request, form)
-            display_message_with_icon(
-                request, messages.INFO, _("The changes have been saved.")
-            )
-            return redirect(reverse("meeting:external_recordings"))
-        else:
-            display_message_with_icon(
-                request,
-                messages.ERROR,
-                _("One or more errors have been found in the form."),
-            )
-
-    page_title = (
-        "%s <b>%s</b>" % (_("Edit the external recording"), recording.name)
-        if recording
-        else _("Add a new external recording")
-    )
-    return render(
-        request,
-        "meeting/add_or_edit_external_recording.html",
-        {"form": form, "page_title": mark_safe(page_title)},
-    )
-
-
-@csrf_protect
-@ensure_csrf_cookie
-@login_required(redirect_field_name="referrer")
-def delete_external_recording(request, id):
-    """Delete an external recording.
-
-    Args:
-        request (Request): HTTP request
-        id (Integer): record id (in database)
-
-    Raises:
-        PermissionDenied: if user not allowed
-
-    Returns:
-        HTTP Response: Redirect to the recordings list
-    """
-    recording = get_object_or_404(Recording, id=id, site=get_current_site(request))
-
-    # Secure external recording
-    secure_external_recording(request, recording)
-
-    # Only POST request
-    secure_post_request(request)
-
-    msg = ""
-    delete = False
-    try:
-        delete = recording.delete()
-    except ValueError as ve:
-        args = ve.args[0]
-        msg = ""
-        if args["error"]:
-            msg += "<strong>%s</strong><br>" % (args["error"])
-        if args["message"]:
-            msg += args["message"]
-    if delete and msg == "":
-        msg += _("The external recording has been deleted.")
-        display_message_with_icon(request, messages.INFO, msg)
-    else:
-        display_message_with_icon(request, messages.ERROR, msg)
-    return redirect(reverse("meeting:external_recordings", args=()))
-
-
-def save_recording_form(request, form):
-    """Save an external recording.
-
-    Args:
-        request (Request): HTTP request
-        form (Form): recording form
-
-    Returns:
-        Recording: recording saved in database
-    """
-    recording = form.save(commit=False)
-    recording.site = get_current_site(request)
-    if (
-        (request.user.is_superuser or request.user.has_perm("meeting.add_recording"))
-        and request.POST.get("owner")
-        and request.POST.get("owner") != ""
-    ):
-        recording.owner = form.cleaned_data["owner"]
-    elif getattr(recording, "owner", None) is None:
-        recording.owner = request.user
-
-    recording.save()
-    form.save_m2m()
-    return recording
-
-
 # ##############################    Upload recordings to Pod
-def parse_remote_file(source_html_url):
-    """Parse the remote HTML file on the BBB server.
-
-    Args:
-        source_html_url (String): URL to parse
-
-    Raises:
-        ValueError: exception raised if no video found in this URL
-
-    Returns:
-        String: name of the video found in the page
-    """
-    try:
-        response = requests.get(source_html_url)
-        if response.status_code != 200:
-            msg = {}
-            msg["error"] = _(
-                "The HTML file for this recording was not found on the BBB server."
-            )
-            # If we want to display the 404/500... page to the user
-            # msg["message"] = response.content.decode("utf-8")
-            msg["message"] = "Error number : %s" % response.status_code
-            raise ValueError(msg)
-
-        # Parse the BBB video HTML file
-        parser = video_parser()
-        # Manage the encoding
-        if response.encoding == "ISO-8859-1":
-            parser.feed(response.text.encode("ISO-8859-1").decode("utf-8"))
-        else:
-            parser.feed(response.text)
-
-        # Video file found
-        if parser.video_check:
-            # Security check about extensions
-            extension = parser.video_file.split(".")[-1].lower()
-            if extension not in VIDEO_ALLOWED_EXTENSIONS:
-                msg = {}
-                msg["error"] = _(
-                    "The video file for this recording was not " "found in the HTML file."
-                )
-                msg["message"] = _("The found file is not a valid video.")
-                raise ValueError(msg)
-
-            # Returns the name of the video (if necessary, title is parser.title)
-            return parser.video_file
-        else:
-            msg = {}
-            msg["error"] = _(
-                "The video file for this recording was not found in the HTML file."
-            )
-            msg["message"] = _("No video file found")
-            raise ValueError(msg)
-    except Exception as exc:
-        msg = {}
-        msg["error"] = _(
-            "The video file for this recording was not found in the HTML file."
-        )
-        msg["message"] = str(exc)
-        raise ValueError(msg)
-
-
-def download_video_file(source_video_url, dest_file):
-    """Download BBB video file.
-
-    Args:
-        source_video_url (String): Video file URL
-        dest_file (String): Destination file of the Pod video
-
-    Raises:
-        ValueError: if impossible download
-    """
-    # Check if video file exists
-    try:
-        with requests.get(source_video_url, timeout=(10, 180), stream=True) as response:
-            if response.status_code != 200:
-                msg = {}
-                msg["error"] = _(
-                    "The video file for this recording "
-                    "was not found on the BBB server."
-                )
-                # If we want to display the 404/500... page to the user
-                # msg["message"] = response.content.decode("utf-8")
-                msg["message"] = "Error number : %s" % response.status_code
-                raise ValueError(msg)
-
-            with open(dest_file, "wb+") as file:
-                # Total size, in bytes, from response header
-                # total_size = int(response.headers.get('content-length', 0))
-                # Other possible methods
-                # Method 1 : iterate over every chunk and calculate % of total
-                # for chunk in response.iter_content(chunk_size=1024*1024):
-                #    file.write(chunk)
-                # Method 2 : Binary download
-                # file.write(response.content)
-                # Method 3 : The fastest
-                shutil.copyfileobj(response.raw, file)
-    except Exception as exc:
-        msg = {}
-        msg["error"] = _("Impossible to download the video file from the server.")
-        msg["message"] = str(exc)
-        raise ValueError(msg)
-
-
-def save_video(request, dest_file, recording_name, description, recording, date_evt=None):
-    """Save and encode the Pod video file.
-
-    Args:
-        request (Request): HTTP request
-        dest_file (String): Destination file of the Pod video
-        recording_name (String): recording name
-        description (String): description added to the Pod video
-        recording (Recording): Recording object
-        date_evt (Datetime, optional): Event date. Defaults to None.
-
-    Raises:
-        ValueError: if impossible creation
-    """
-    try:
-        video = Video.objects.create(
-            video=dest_file,
-            title=recording_name,
-            owner=request.user,
-            description=description,
-            is_draft=True,
-            type=Type.objects.get(id=DEFAULT_TYPE_ID),
-            date_evt=date_evt,
-        )
-        for additional_owner in recording.additional_owners.all():
-            video.additional_owners.add(additional_owner)
-
-        video.launch_encode = True
-        video.save()
-    except Exception as exc:
-        msg = {}
-        msg["error"] = _("Impossible to create the Pod video")
-        msg["message"] = str(exc)
-        raise ValueError(msg)
-
-
 def save_internal_recording(
     request, recording_id, recording_name, meeting_id, source_url=None
 ):
@@ -1514,9 +1034,8 @@ def save_internal_recording(
         if source_url is None:
             source_url = ""
         # Save the recording as an internal recording
-        recording, created = Recording.objects.update_or_create(
+        recording, created = InternalRecording.objects.update_or_create(
             name=recording_name,
-            is_internal=True,
             recording_id=recording_id,
             meeting=meeting,
             # start_at=start_at,
@@ -1527,30 +1046,6 @@ def save_internal_recording(
     except Exception as exc:
         msg = {}
         msg["error"] = _("Impossible to create the internal recording")
-        msg["message"] = str(exc)
-        raise ValueError(msg)
-
-
-def save_external_recording(request, record_id):
-    """Save an external recording in database.
-
-    Args:
-        request (Request): HTTP request
-        record_id (Integer): id of the recording in database
-
-    Raises:
-        ValueError: if impossible creation
-    """
-    try:
-        # Update the external recording
-        recording, created = Recording.objects.update_or_create(
-            id=record_id,
-            is_internal=False,
-            defaults={"uploaded_to_pod_by": request.user},
-        )
-    except Exception as exc:
-        msg = {}
-        msg["error"] = _("Impossible to create the external recording")
         msg["message"] = str(exc)
         raise ValueError(msg)
 
@@ -1570,19 +1065,10 @@ def upload_recording_to_pod(request, record_id, meeting_id=None):
         Boolean: True if upload achieved
     """
     try:
-        # Management by type of recording
-        recording = Recording.objects.get(id=record_id)
-
         # Check that request is correct for upload
         secure_request_for_upload(request)
 
-        # Manage differents source types
-        if recording.type == "youtube":
-            return upload_youtube_recording_to_pod(request, record_id)
-        elif recording.type == "peertube":
-            return upload_peertube_recording_to_pod(request, record_id)
-        else:
-            return upload_bbb_recording_to_pod(request, record_id, meeting_id)
+        return upload_bbb_recording_to_pod(request, record_id, meeting_id)
     except Exception as exc:
         msg = {}
         proposition = ""
@@ -1600,7 +1086,7 @@ def upload_recording_to_pod(request, record_id, meeting_id=None):
         raise ValueError(msg)
 
 
-def upload_bbb_recording_to_pod(request, record_id, meeting_id=None):
+def upload_bbb_recording_to_pod(request, record_id, meeting_id):
     """Upload a BBB or video file recording to Pod.
 
     Args:
@@ -1615,7 +1101,7 @@ def upload_bbb_recording_to_pod(request, record_id, meeting_id=None):
         Boolean: True if upload achieved
     """
     try:
-        recording = Recording.objects.get(id=record_id)
+        recording = InternalRecording.objects.get(id=record_id)
         source_url = request.POST.get("source_url")
 
         # Step 1 : Download and parse the remote HTML file if necessary
@@ -1646,24 +1132,21 @@ def upload_bbb_recording_to_pod(request, record_id, meeting_id=None):
 
         # Step 4 : Save informations about the recording
         recording_title = request.POST.get("recording_name")
-        if meeting_id is not None:
-            save_internal_recording(
-                request,
-                recording.recording_id,
-                recording_title,
-                meeting_id,
-                source_video_url,
-            )
-        else:
-            save_external_recording(request, record_id)
+        save_internal_recording(
+            request,
+            recording.recording_id,
+            recording_title,
+            meeting_id,
+            source_video_url,
+        )
 
         # Step 5 : Save and encode Pod video
         description = _(
             "This video was uploaded to Pod; its origin is %(type)s : "
             '<a href="%(url)s" target="_blank">%(url)s</a>'
-        ) % {"type": recording.get_type_display(), "url": source_video_url}
+        ) % {"type": "Big Blue Button", "url": source_video_url}
 
-        save_video(request, dest_file, recording_title, description, recording)
+        save_video(request, dest_file, recording_title, description)
 
         return True
     except Exception as exc:
@@ -1681,277 +1164,3 @@ def upload_bbb_recording_to_pod(request, record_id, meeting_id=None):
             "Try changing the record type or address for this recording."
         )
         raise ValueError(msg)
-
-
-def upload_youtube_recording_to_pod(request, record_id):
-    """Upload Youtube recording to Pod.
-
-    Use PyTube with its API
-    More information : https://pytube.io/en/latest/api.html
-    Args:
-        request (Request): HTTP request
-        record_id (Integer): id record in the database
-
-    Raises:
-        ValueError: exception raised if no YouTube video found or content inaccessible
-
-    Returns:
-        Boolean: True if upload achieved
-    """
-    try:
-        recording = Recording.objects.get(id=record_id)
-
-        # Manage source URL from video playback
-        source_url = request.POST.get("source_url")
-
-        # Use pytube to download Youtube file
-        yt_video = YouTube(
-            source_url,
-            # on_complete_callback=complete_func,
-            # use_oauth=True,
-            # allow_oauth_cache=True
-        )
-        # Publish date (format : 2023-05-13 00:00:00)
-        # Event date (format : 2023-05-13)
-        date_evt = str(yt_video.publish_date)[0:10]
-
-        # Setting video resolution
-        yt_stream = yt_video.streams.get_highest_resolution()
-
-        # User directory
-        dest_dir = os.path.join(
-            settings.MEDIA_ROOT,
-            "videos",
-            request.user.owner.hashkey,
-        )
-        os.makedirs(os.path.dirname(dest_dir), exist_ok=True)
-
-        discrim = datetime.now().strftime("%Y%m%d%H%M%S")
-        filename = "%s-%s" % (discrim, yt_stream.default_filename)
-        # Video file path
-        dest_file = os.path.join(
-            dest_dir,
-            filename,
-        )
-
-        # Download video
-        yt_stream.download(dest_dir, filename=filename)
-
-        # Step 4 : Save informations about the recording
-        save_external_recording(request, record_id)
-
-        # Step 5 : Save and encode Pod video
-        description = _(
-            "This video '%(name)s' was uploaded to Pod; "
-            'its origin is Youtube : <a href="%(url)s" target="_blank">%(url)s</a>'
-        ) % {"name": yt_video.title, "url": source_url}
-        recording_title = request.POST.get("recording_name")
-        save_video(request, dest_file, recording_title, description, recording, date_evt)
-        return True
-
-    except VideoUnavailable:
-        msg = {}
-        msg["error"] = _("YouTube error")
-        msg["message"] = _(
-            "YouTube content is unavailable. "
-            "This content does not appear to be publicly available."
-        )
-        msg["proposition"] = _(
-            "Try changing the access rights to the video directly in Youtube."
-        )
-        raise ValueError(msg)
-    except PytubeError:
-        msg = {}
-        msg["error"] = _("YouTube error")
-        msg["message"] = _(
-            "YouTube content is inaccessible. "
-            "This content does not appear to be publicly available."
-        )
-        msg["proposition"] = _("Try changing the address of this recording.")
-        raise ValueError(msg)
-    except Exception as exc:
-        msg = {}
-        msg["error"] = _("Impossible to upload to Pod the video")
-        try:
-            # Management of error messages from sub-functions
-            message = "%s (%s)" % (exc.args[0]["error"], exc.args[0]["message"])
-        except Exception:
-            # Management of error messages in all cases
-            message = str(exc)
-
-        msg["message"] = mark_safe(message)
-        msg["proposition"] = _(
-            "Try changing the record type or address for this recording."
-        )
-        raise ValueError(msg)
-
-
-def upload_peertube_recording_to_pod(request, record_id):  # noqa: C901
-    """Upload Peertube recording to Pod.
-
-    More information : https://docs.joinpeertube.org/api/rest-getting-started
-    Args:
-        request (Request): HTTP request
-        record_id (Integer): id record in the database
-
-    Raises:
-        ValueError: exception raised if no PeerTube video found in this URL
-
-    Returns:
-        Boolean: True if upload achieved
-    """
-    try:
-        recording = Recording.objects.get(id=record_id)
-
-        # Manage source URL from video playback
-        source_url = request.POST.get("source_url")
-
-        # Check if extension is a video extension
-        extension = source_url.split(".")[-1].lower()
-        if extension in VIDEO_ALLOWED_EXTENSIONS:
-            # URL corresponds to a video file. Format example :
-            #  - https://xxxx.fr/download/videos/id-quality.mp4
-            # with : id = id/uuid/shortUUID, quality=480/720/1080
-            source_video_url = source_url
-            # PeerTube API for this video :
-            # https://xxxx.fr/api/v1/videos/id
-            pos_pt = source_url.rfind("-")
-            if pos_pt != -1:
-                url_api_video = source_url[0:pos_pt].replace(
-                    "/download/videos/", "/api/v1/videos/"
-                )
-            else:
-                msg = {}
-                msg["error"] = _("PeerTube error")
-                msg["message"] = _(
-                    "The address entered does not appear to be a valid PeerTube address."
-                )
-                msg["proposition"] = _("Try changing the address of the recording.")
-                raise ValueError(msg)
-        else:
-            # URL corresponds to a PeerTube URL. Format example :
-            #  - https://xxx.fr/w/id
-            #  - https://xxx.fr/videos/watch/id
-            # with : id = id/uuid/shortUUID
-            # PeerTube API for this video :
-            # https://xxxx.fr/api/v1/videos/id
-            url_api_video = source_url.replace("/w/", "/api/v1/videos/")
-            url_api_video = url_api_video.replace("/videos/watch/", "/api/v1/videos/")
-
-        with requests.get(url_api_video, timeout=(10, 180), stream=True) as response:
-            if response.status_code != 200:
-                msg = {}
-                msg["error"] = _("PeerTube error")
-                msg["message"] = _(
-                    "The address entered does not appear to be a valid PeerTube address "
-                    "or the PeerTube server is not responding as expected."
-                )
-                msg["proposition"] = _(
-                    "Try changing the address of the recording or retry later."
-                )
-                raise ValueError(msg)
-            else:
-                pt_video_json = json.loads(response.content.decode("utf-8"))
-                # URL
-                pt_video_url = pt_video_json["url"]
-                # UUID, useful for the filename
-                pt_video_uuid = pt_video_json["uuid"]
-                pt_video_name = pt_video_json["name"]
-                pt_video_description = pt_video_json["description"]
-                if pt_video_description is None:
-                    pt_video_description = ""
-                else:
-                    pt_video_description = pt_video_description.replace("\r\n", "<br>")
-                # Creation date (format : 2023-05-23T08:16:34.690Z)
-                pt_video_created_at = pt_video_json["createdAt"]
-                # Evant date (format : 2023-05-23)
-                date_evt = pt_video_created_at[0:10]
-                # Source video file
-                source_video_url = pt_video_json["files"][0]["fileDownloadUrl"]
-
-        # Step 2 : Define destination source file
-        discrim = datetime.now().strftime("%Y%m%d%H%M%S")
-        extension = source_video_url.split(".")[-1].lower()
-        dest_file = os.path.join(
-            settings.MEDIA_ROOT,
-            "videos",
-            request.user.owner.hashkey,
-            os.path.basename("%s-%s.%s" % (discrim, pt_video_uuid, extension)),
-        )
-        os.makedirs(os.path.dirname(dest_file), exist_ok=True)
-
-        # Step 3 : Download the video file
-        download_video_file(source_video_url, dest_file)
-
-        # Step 4 : Save informations about the recording
-        recording_title = request.POST.get("recording_name")
-        save_external_recording(request, record_id)
-
-        # Step 5 : Save and encode Pod video
-        description = _(
-            "This video '%(name)s' was uploaded to Pod; its origin is PeerTube : "
-            "<a href='%(url)s' target='blank'>%(url)s</a>."
-        ) % {"name": pt_video_name, "url": pt_video_url}
-        description = ("%s<br>%s") % (description, pt_video_description)
-        save_video(request, dest_file, recording_title, description, recording, date_evt)
-
-        return True
-    except Exception as exc:
-        msg = {}
-        msg["error"] = _("Impossible to upload to Pod the PeerTube video")
-        try:
-            # Management of error messages from sub-functions
-            message = "%s (%s)" % (exc.args[0]["error"], exc.args[0]["message"])
-        except Exception:
-            # Management of error messages in all cases
-            message = str(exc)
-
-        msg["message"] = mark_safe(message)
-        msg["proposition"] = _(
-            "Try changing the record type or address for this recording."
-        )
-        raise ValueError(msg)
-
-
-class video_parser(HTMLParser):
-    """Useful to parse the BBB Web page and search for video file.
-
-    Used to parse BBB 2.6+ URL for video recordings.
-    Args:
-        HTMLParser (_type_): _description_
-    """
-
-    def __init__(self):
-        """Initialize video parser."""
-        super().__init__()
-        self.reset()
-        # Variables for title
-        self.title_check = False
-        self.title = ""
-        # Variables for video file
-        self.video_check = False
-        self.video_file = ""
-        self.video_type = ""
-
-    def handle_starttag(self, tag, attrs):
-        """Parse BBB Web page and search video file."""
-        attrs = dict(attrs)
-        # Search for source tag
-        if tag == "source":
-            # Found the line. Managed format :
-            # attrs = {'src': 'video-0.m4v', 'type': 'video/mp4'}
-            # print("video line : %s" % attrs)
-            self.video_check = True
-            self.video_file = attrs.get("src", "")
-            self.video_type = attrs.get("type", "")
-        # Search for title tag
-        if tag == "title":
-            # Found the title line
-            self.title_check = True
-
-    def handle_data(self, data):
-        """Search for title tag."""
-        if self.title_check:
-            # Get the title that corresponds to recording's name
-            self.title = data
-            self.title_check = False
