+++ requirements-dev.txt
@@ -1,3 +1,4 @@
 flake8
 coveralls
 httmock
+beautifulsoup4
